fix: refactor rake rollbar:test to not autoload during app init
